move normalizing qr ticket id to component itself
